requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic-settings==2.7.1
httpx==0.28.1

# LLM Providers
openai==1.59.7
anthropic==0.42.0

# PDF Processing
PyPDF2==3.0.1

# Embeddings & Vector Store
sentence-transformers==3.3.1
faiss-cpu==1.9.0.post1
chromadb==0.5.23
numpy>=1.26.0

# Resilience
tenacity==9.0.0

# Testing
pytest==8.3.4
